update to be compatible with ros_type_introspection 0.6

+++ plugins/ROS/DataLoadROS/dataload_ros.cpp
@@ -180,7 +180,7 @@ PlotDataMap DataLoadROS::readDataFromFile(const QString &file_name,
                 }
 
                 PlotDataPtr& plot_data = plot_pair->second;
-                plot_data->pushBack( PlotData::Point(msg_time, it.second));
+                plot_data->pushBack( PlotData::Point(msg_time, (double)it.second));
             } //end of for flat_container.renamed_value
 
             //-----------------------------------------
